`SCRUM-40` game modal preps
