--- packages/client/src/components/FinishModal/FinishModal.module.scss
@@ -1,18 +1,0 @@
-@import '@styles/vars.scss';
-@import '@styles/mixins.scss';
-
-.wrapper {
-  @include center-flex;
-}
-
-.modal {
-  width: 400px;
-  min-height: 250px;
-  background-color: $color-bg-secondary;
-  border-radius: 12px;
-  padding: 24px;
-  outline: none;
-  display: flex;
-  flex-direction: column;
-  gap: 16px
-}
