--- packages/client/src/components/FinishModal/index.ts
@@ -1,2 +1,0 @@
-export * from './FinishModal'
-export * from './types'
