--- packages/client/src/components/FinishModal/types.ts
@@ -1,1 +1,0 @@
-export type TFinishMode = 'win' | 'lose'
