+++ packages/client/src/components/GameModal/index.ts
@@ -1,0 +1,2 @@
+export * from './GameModal'
+export * from './types'
