+++ packages/client/src/components/GameModal/types.ts
@@ -1,0 +1,1 @@
+export type TGameModalMode = 'closed' | 'start' | 'pause' | 'win' | 'lose'
