--- packages/client/src/components/PreviewModal/index.ts
@@ -1,2 +1,0 @@
-export * from './PreviewModal'
-export * from './types'
