--- packages/client/src/components/PreviewModal/types.ts
@@ -1,1 +1,0 @@
-export type TOpenMode = 'start' | 'pause'
