+++ packages/client/src/components/index.ts
@@ -2,5 +2,4 @@ export * from './Form'
 export * from './FullScreenWrapper'
 export * from './Input'
 export * from './FormField'
-export * from './PreviewModal'
-export * from './FinishModal'
+export * from './GameModal'
